Provide an API for eager and complete diagnostics.
This change introduces the `DiagnosticGlob` trait for constructing globs
with complete diagnostics (such that error, warnings, etc. are detected
and reported together as comprehensively as possible).

To faciliate this, this change also introduces nuance in token APIs to
better reflect the relationship between referenced inputs and the
originating glob expression (typically denoted with the lifetime `'t`).